TY: first step to the full type inference
