+++ src/main/kotlin/org/rust/lang/core/resolve/ref/RsFieldExprReferenceImpl.kt
@@ -8,8 +8,12 @@ package org.rust.lang.core.resolve.ref
 import com.intellij.psi.PsiElement
 import org.rust.lang.core.psi.RsFieldExpr
 import org.rust.lang.core.psi.ext.RsCompositeElement
-import org.rust.lang.core.resolve.*
+import org.rust.lang.core.resolve.collectCompletionVariants
+import org.rust.lang.core.resolve.collectResolveVariants
+import org.rust.lang.core.resolve.processFieldExprResolveVariants
 import org.rust.lang.core.types.BoundElement
+import org.rust.lang.core.types.ty.Ty
+import org.rust.lang.core.types.type
 
 class RsFieldExprReferenceImpl(
     fieldExpr: RsFieldExpr
@@ -20,16 +24,23 @@ class RsFieldExprReferenceImpl(
 
     override fun getVariants(): Array<out Any> =
         collectCompletionVariants {
-            processFieldExprResolveVariants(element, true, it)
+            processFieldExprResolveVariants(element.project, element.expr.type, true, it)
         }
 
     override fun resolveInner(): List<BoundElement<RsCompositeElement>> =
-        collectResolveVariants(element.referenceName) {
-            processFieldExprResolveVariants(element, false, it)
-        }
+        resolveFieldExprReferenceWithReceiverType(element.expr.type, element)
 
     override fun handleElementRename(newName: String): PsiElement {
         element.fieldId.identifier?.let { doRename(it, newName) }
         return element
+    }
+}
+
+fun resolveFieldExprReferenceWithReceiverType(
+    receiverType: Ty,
+    expr: RsFieldExpr
+): List<BoundElement<RsCompositeElement>> {
+    return collectResolveVariants(expr.referenceName) {
+        processFieldExprResolveVariants(expr.project, receiverType, false, it)
     }
 }
